C files singly linked lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -9,7 +9,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned long i = 0;
+	size_t i = 0;
 	const listint_t *current_node;
 
 	if (h == NULL)
